Reorganize for testing
