+++ .gitignore
@@ -10,5 +10,3 @@ logs
 *.map
 dist
 distribution
-index.js
-index.d.ts
